% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrep.R
\name{dataPrep}
\alias{dataPrep}
\title{Read in a dataset and prepare it for analysis}
\usage{
dataPrep(
  file.name,
  time = "Time",
  beads = c("Bead"),
  dna = c("DNA1", "DNA2"),
  event_length = "Event_length",
  viability = "Live_Dead",
  gaussian = c("Center", "Offset", "Width", "Residual")
)
}
\arguments{
\item{file.name}{A path to an .fcs file that contains CyTOF data.}

\item{beads}{character vector that contains the names of all of the bead channels.}

\item{dna}{Character vector that contains the names of the DNA markers.}

\item{event_length}{Character vector of the event length variable.}

\item{viability}{Character vector of the permeability/viability markers.}
}
\value{
A \code{matrix} that contains the time, beads, DNA measures, permeability
(viability), event length, and the Guassian parameters of the CyTOF data.
Note that the time variable is included so that the cells can be matched up
exactly with the QC label data at any point in the process.
}
\description{
Read in a dataset and prepare it for analysis
}
\details{
The function returns a \code{matrix} that contains the beads, DNA measures,
event length, permeability (viability) measure, and Gaussian parameters.
The event length and Gaussian parameters are transformed with a natural
log transformation. The data are not standardized at this point because
the raw data are needed for some of the data labeling steps.
}
\examples{
fname <- "../data/FlowRepository_FR-FCM-Z29V_files/REP_1_deid.fcs"

# See names of variables
require(CATALYST)
require(SingleCellExperiment)
sce <- prepData(fname)
rownames(sce)
rowData(sce)
names(int_colData(sce))

# Read in data
tech <- dataPrep(fname)
}
