% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_groups.R
\name{find_groups}
\alias{find_groups}
\title{splits a numeric vector into groups.}
\usage{
find_groups(x)
}
\arguments{
\item{x}{A numeric vector to be split into groups.}
}
\value{
A vector that contains the groups assignment for each element
of \code{x}.
}
\description{
splits a numeric vector into groups.
}
\details{
The function determines how many peaks are in the data and then
splits the data into groups that belong to each peak.
}
\examples{
data("raw_data", package = "CATALYST")
tech <- dataPrep(raw_data, beads = 'Beads', viability = c('cisPt1','cisPt2'))
find_groups(tech[,"Bead1"])

}
