% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate a cytofQC report}
\usage{
generateReport(tech, labels, outDir, sampName, runUMAP = TRUE, ...)
}
\arguments{
\item{tech}{A matrix of technical variables, as generated by
\code{\link{dataPrep}}}

\item{labels}{A \code{data.frame} containing predicted event labels, as
generated by \code{\link{labelQC}}}

\item{outDir}{The output directory (currently required).}

\item{sampName}{Basename of the output HTML file (if not provided, same as
\code{outDir}).}

\item{runUMAP}{Logical value indicating whether or not to include a UMAP plot
in the report. This plot can be beneficial for diagnostic purposes, but is
time-consuming to generate.}

\item{...}{Additional arguments passed to \code{\link{labelQC}}, if
\code{labels} was not provided.}
}
\value{
If successful, returns \code{TRUE} silently and generates the
  specified QC report.
}
\description{
Generate a cytofQC report
}
\examples{
fname <- system.file("extdata", "raw_cytof.fcs", package = "cytofQC")
tech <- dataPrep(fname)
labels <- labelQC(tech, nTrain = 1000)
tmp <- tempdir()
generateReport(tech, labels, tmp, 'example')

}
