% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_score}
\alias{get_score}
\alias{get_prob}
\alias{get_labels}
\alias{get_tech}
\title{Returns a specified object from the cytofQC SingleCellExperiment}
\usage{
get_score(x, type = NULL)

get_prob(x, type = NULL)

get_labels(x)

get_tech(x)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} created with \code{\link{readCytof}} 
with the scores and initial columns filled out for the event type of 
interest.}

\item{type}{Identifies the type of object to be returned. for 
\code{get_score} and \code{get_prob}, type can be 'bead', 'dead', 'debris', 
or 'doublet' and it will return the numeric vector for the score or 
probability for the specified event type. If it is not specified, 
the \code{DataFrame} containing all of the scores or all of the 
probability will be returned. This argument does nothing for 
\code{get_labels} and \code{get_tech}.}
}
\value{
For \code{get_prob} and \code{get_score}, a numeric vector 
containing the score or probabilities for the event type is returned. 
If \code{type} is not specified, a \code{DataFrame} containing all 
of the probabilities or scores is returned. 

For \code{get_labels}, a character vector containing the label
for each event is returned.

For \code{get_tech}, a \code{DataFrame} containing the technical variables
used to determine the label of each event. The bead, DNA, and viability
variables have an arcsinh transform, Event_length is unchanged, and 
the Gaussian parameters have a log transform using \code{log1p}.
}
\description{
Returns a specified object from the cytofQC SingleCellExperiment
}
\examples{
data("raw_data", package = "CATALYST")
sce <- readCytof(raw_data, beads = "Beads", viability = c("cisPt1", "cisPt2"))
sce <- initialBead(sce)
sce <- svmLabel(sce, type = "bead", loss = "auc")
cytofHist(get_score(sce, type = 'bead'), get_labels(sce), title = "Bead score")
}
