% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.R
\name{modelData}
\alias{modelData}
\title{Returns indices for data to be used to create the final classification model.}
\usage{
modelData(labels, init, subset = NULL, n = 4000)
}
\arguments{
\item{labels}{A \code{data.frame} created with \code{\link{qcDataFrame}}.}

\item{init}{A numeric vector that contains the initial labeling for the
observations for the parameter of interest.}

\item{subset}{A logical vector that indicates if an observation should
be considered for inclusion in the returned indices. If there are fewer 
TRUE values in \code{subset} than \code{n}, all of the unclassified data 
will be used as the subset.}

\item{n}{number of indices to return.}
}
\value{
An integer vector that contains the indices of the events that
should be included in the creation of the final classification model for
the parameter of interest (bead, debris, doublet, dead).
}
\description{
Returns indices for data to be used to create the final classification model.
}
\details{
The indices that are returned by \code{modelData} should be used to
create a model that can be used to classify the observations with
regard to the parameter of interest (bead, doublet, debris, dead).
}
\examples{
fname <- "../data/FlowRepository_FR-FCM-Z29V_files/REP_1_deid.fcs"
x <- dataPrep(fname)
nn <- NN(x)
labels <- qcDataFrame(x)
beads <- beadID(x, labels = labels)
index <- modelData(labels, init = beads$init, n = 4000)

}
