% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmLabel.R
\name{svmLabel}
\alias{svmLabel}
\title{Returns the final label assignments for a parameter using a support vector
machine}
\usage{
svmLabel(x, labels, type, init, index, loss = "auc", standardize = TRUE)
}
\arguments{
\item{x}{A \code{matrix} created with \code{\link{dataPrep}}.}

\item{labels}{A \code{data.frame} created with \code{\link{qcDataFrame}}.}

\item{type}{Identifies the type of label that is being modeled. Must
be 'bead', 'doublet', 'debris', or 'dead'.}

\item{init}{A logical vector that contains the initial labeling for the
cells for the cell type of interest.}

\item{index}{A vector containing the indices of the data that should be
used to compute the model. These should be obtained from
\code{\link{modelData}}.}

\item{loss}{Specifies the type of loss used to tune the GBM. Can be either
"auc" or "class".}

\item{standardize}{Indicates if the data should be standardized. Because
the data are on different scales, it should be standardized for
this analysis.}
}
\value{
An updated \code{label} data.frame is returned with the labels
for the parameter of interest (bead, doublet, debris, or dead) added to
the \code{label} variable and the probabilities for the column
pertaining to the parameter filled in.
}
\description{
Returns the final label assignments for a parameter using a support vector
machine
}
\details{
\code{svmLabel} uses a support vector machine to compute the final
labels for the specified parameter type (bead, doublet, debris, or dead). The
model is computed using only the data specified in the index argument. The
svm is tuned using \code{\link[EZtune]{eztune}} and then predicted values are
computed for all of the observations in \code{x}. If the predicted
probability for the label type is greater than 0.5, the label is changed to
the specified type. However, if an observation already has a label other than
'cell' in the \code{labels$label} variable, it will not be changed. The
predicted probabilities for all of the observations is stored in the variable
associated with that type for further analysis. Thus, it is possible to have
a probability greater than 0.5 for 'debris' but still have a label of 'bead'
if an observation was classified as a bead prior to classifying the debris.
}
\examples{
data("raw_data", package = "CATALYST")
tech <- dataPrep(raw_data, beads = 'Beads', viability = c('cisPt1','cisPt2'))
lab <- qcDataFrame(tech)
beads <- initialBead(tech, lab)
sure <- beads$init \%in\% c(-1,1)
ind <- modelData(lab, subset = sure, init = beads$init)
svmLabel(tech, lab, type = "bead", init = beads$init, index = ind)

}
