% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelQC.R
\name{labelQC}
\alias{labelQC}
\title{Returns the final label assignments the specified parameters}
\usage{
labelQC(
  x,
  model = "svm",
  types = c("bead", "doublet", "debris", "dead"),
  loss = "auc"
)
}
\arguments{
\item{x}{A \code{matrix} created with \code{\link{dataPrep}}.}

\item{model}{Type of model to use to do the labeling. Options are 
"svm" for a support vector machine, "gbm" for a gradient boosting
machine, or "rf" for a random forest.}

\item{types}{Types of to model. Options are "bead", "doublet", 
"debris", and "dead".}

\item{loss}{Specifies the type of loss used to tune the model. Can 
be either "auc" or "class". This argument is ignored if random
forest is used as the model.}
}
\value{
A \code{label} data.frame is returned with the labels
for the parameters of listed in \code{types} (bead, doublet, debris, 
or dead) added to the \code{label} variable and the probabilities for 
each of the columns pertaining to the parameter listed in \code{types}.
}
\description{
Returns the final label assignments the specified parameters
}
\details{
\code{labelQC} uses a support vector machine, gradient boosting machine, 
or a random forest to compute the final labels
for the specified parameter types (bead, doublet, debris, or dead). 
The predicted probabilities for all of the observations are stored in 
the variable associated with that type for further analysis. Thus, it 
is possible to have a probability greater than 0.5 for 'debris' but still 
have a label of 'bead' if an observation was classified as a bead prior to
classifying the debris.
}
\examples{
fname <- "../data/FlowRepository_FR-FCM-Z29V_files/REP_1_deid.fcs"
x <- dataPrep(fname)
svmLabels <- labelQC(x, model = "svm", types = c("bead", "doublet", "debris"))

}
