% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInitialGuess.R
\name{plotInitialGuess}
\alias{plotInitialGuess}
\title{Plot preliminary classification from initialGuess.}
\usage{
plotInitialGuess(x, IG = NULL, fit = NULL)
}
\arguments{
\item{x}{The score (ie. debris score, doublet score, etc.) to be used for
predicting each event's label (eg. "doublet" vs. "cell").}

\item{IG}{If NULL, the function \code{initialGuess} is used to fit a mixture
of normal distributions. Otherwise, a numeric vector can be passed to the 
function that contains the fitted values.}

\item{fit}{If left blank or NULL, the best fit as determined by BIC will be
plotted. Otherwise, a numeric value of 1, 2, or 3 can be be selected to 
plot single normal fit, mixture of two normals, or the mixture of three
normals as fit by \code{initialGuess}.}
}
\value{
A list with the following elements: \itemize{ \item{\code{label}} {A
  vector of the same length as \code{x} providing the labels (\code{-1} for
  cells, \code{1} for non-cells, \code{0} for uncertain).} \item{\code{fit1}}
  {Summary of the 1-component (half Normal) model fit.} \item{\code{fit2}}
  {Summary of the 2-component (half Normal + Normal) model fit.}
  \item{\code{fit1}} {Summary of the 3-component (half Normal + 2 Normals)
  model fit.}}
}
\description{
Plot preliminary classification from initialGuess.
}
\examples{
data("raw_data", package = "CATALYST")
tech <- dataPrep(raw_data, beads = 'Beads', viability = c('cisPt1','cisPt2'))
lab <- qcDataFrame(tech)
doublets <- initialDoublet(tech, labels = lab, score = 1)
cytofQC:::initialGuess(doublets$doubletScore)

}
