% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfLabel.R
\name{rfLabel}
\alias{rfLabel}
\title{Returns the final label assignments for a parameter using a random
forest}
\usage{
rfLabel(
  x,
  type = c("bead", "doublet", "debris", "dead"),
  loss = "auc",
  n = 4000,
  standardize = TRUE
)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} created with \code{\link{readCytof}} 
with the scores and initial columns filled out for the event type of 
interest.}

\item{type}{Identifies the type of label that is being modeled. Must
be 'bead', 'doublet', 'debris', or 'dead'.}

\item{loss}{Specifies the type of loss used to tune the GBM. Can be either
"auc" for the area under the curve or "class" for classification error.}

\item{n}{number of observations in training dataset.}

\item{standardize}{Indicates if the data should be standardized. Because
the data are on different scales, it should be standardized for
this analysis because the variables are on different scales.}
}
\value{
An updated \code{SingleCellExperiment} is returned with the labels
for the parameter of interest (bead, doublet, debris, or dead) added to
the \code{label} object of the \code{SingleCellExperiment} and the 
probabilities for the event type added to the \code{probs} object of the 
\code{SingleCellExperiment}.
}
\description{
Returns the final label assignments for a parameter using a random
forest
}
\details{
\code{rfLabel} uses a random forest to compute the final labels
for the specified parameter type (bead, doublet, debris, or dead). This step
cannot be completed until the corresponding initialization function 
(\code{initialBead}, \code{initialDebris}, \code{initialDoublet}, or 
\code{initialDead}) is done on the \code{SingleCellExperiment} created by 
\code{readCytof}. 
The random forest uses the defaults from \code{randomForest} and then predicted 
values are computed for all of the events in \code{x}. If the predicted
probability for the label type is greater than 0.5, the label is changed to
the specified type. However, if an observation already has a label other 
than 'cell' in the \code{label} variable, it will not be changed. The 
predicted probabilities for all of the observations are stored in the 
variable associated with that type in the \code{probs} object of \code{x}
for further analysis. Thus, it is possible to have a probability greater 
than 0.5 for 'debris' but still have a label of 'bead' if an observation 
was classified as a bead prior to classifying the debris.
}
\examples{
data("raw_data", package = "CATALYST")
sce <- readCytof(raw_data, beads = "Beads", viability = c("cisPt1", "cisPt2"))
sce <- initialBead(sce)
sce <- rfLabel(sce, type = "bead")

}
